Make `Prop` not generic
Caused more trouble than it was worth with this implementation. Might be
worth it to revisit in the future.

+++ article-api/src/test/scala/no/ndla/articleapi/repository/ArticleRepositoryTest.scala
@@ -31,7 +31,7 @@ class ArticleRepositoryTest
   lazy val sampleArticle: Article = TestData.sampleArticleWithByNcSa
 
   def serverIsListening: Boolean = {
-    Try(new Socket(props.MetaServer.unsafeGet, props.MetaPort.unsafeGet)) match {
+    Try(new Socket(props.MetaServer.unsafeGet, props.MetaPort.unsafeGet.toInt)) match {
       case Success(c) =>
         c.close()
         true
