+++ concept-api/src/test/scala/no/ndla/conceptapi/repository/DraftConceptRepositoryTest.scala
@@ -51,7 +51,7 @@ class DraftConceptRepositoryTest
   }
 
   def serverIsListening: Boolean = {
-    Try(new Socket(props.MetaServer.unsafeGet, props.MetaPort.unsafeGet)) match {
+    Try(new Socket(props.MetaServer.unsafeGet, props.MetaPort.unsafeGet.toInt)) match {
       case Success(c) =>
         c.close()
         true
